requirements.txt
Django==3.1.3
djangorestframework==3.12.2
djangorestframework-simplejwt
drf-yasg==1.20.0
mysqlclient==2.0.3
django-cors-headers==3.7.0
